some renaming + dynamic year + component created for timer in appbar (avoiding code duplications)
