+++ .hz/config.toml
@@ -25,7 +25,7 @@ port = 8080
 #                application state
 # 'serve_static' will serve files from the given directory over HTTP/HTTPS
 #------------------------------------------------------------------------------
-project_name = "bouffe_2016"
+project_name = "ua_bouffe"
 # serve_static = "dist"
