+++ src/components/Index.jsx
@@ -11,10 +11,12 @@ export default class Index extends React.Component {
   }
 
   render() {
+    const year = new Date().getFullYear();
+
     return (
       <div className="b-chooser">
         <div className="b-chooser__title">
-          <span>UA Bouffe 2016</span>
+          <span>UA {year} - Bouffe</span>
         </div>
         <div className="b-chooser__link b-chooser__link--seller" onClick={() => this.sell()}>
           Interface de vente
